feat: specify version to install and install path
This also reverts the last commit and we are no longer using the cached
manifest.

The reason for this is that indieGala's servers reject downloads of
chunks for older "disabled" builds unless you have requested the old
version manifest before attempting to download the "disabled" chunk.

As there is little benefit to over-complicate the logic for whether to
use the cache or not based on if the version is the latest or manually
selected, we just stick to always downloading the latest manifests for
the version we want to install.