feat: memory limit and updating games
Downloads should now be faster as we are unlocking new download threads
as long as the memory buffer for chunks is not full, instead of only
when chunks are flushed to disk, which limited our download tasks to how
fast we could download and write `max-download-workers` sequential chunks
to disk.

Updates should now be working as well. Theoretically this logic should
also work for downgrading games in the future.