fix: handle chunk SHA not containing ID
Die Young: Prologue suffers from this issue. The Chunk SHA in it's
manifest does not follow the format
{FILENAME_MD5}_{CHUNK_IDX}_{CHUNK_SHA}, but the format
{FILENAME_MD5}_{CHUNK_SHA}.

Also, since it doesn't contain the chunk index, it's no longer
guaranteed to be unique, since there are chunks that are exactly the
same in the same file. To fix this we are prepending the chunk index
manually in the input buffer to ensure that each key is unique.